graph: make fdpclust example a doc example
